requirements.txt
protobuf==4.25.1
Pillow==10.1.0
numpy==1.24.3
